python_a2a_demo/pyproject.toml
[project]
name = "python_a2a_demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "numexpr>=2.11.0",
    "python-a2a[all]>=0.5.9",
    "python-dotenv>=1.1.1",
]
